+++ CHANGELOG.md
@@ -10,6 +10,14 @@ Some smaller breaking changes, mostly polish around public APIs
 
 * The simulator is now [available on crates.io](https://crates.io/crates/embedded-graphics-simulator) as a standalone crate. You can now create simulated displays for testing out embedded_graphics code or showing off cool examples.
 
+### Changed
+
+* **(breaking)** Primitives macros have been renamed. This is primarily to fix conflicts with `std`'s `line!()` macro, but I thought I'd take the opportunity to make the names a bit better/more consistent at the same time:
+  * `line` -> `egline`
+  * `triangle` -> `egtriangle`
+  * `rect` -> `egrectangle`
+  * `circle` -> `egcircle`
+
 ## 0.5.0
 
 A big release, focussed on ergonomics. There are new macros to make drawing and positioning primitives and text much less noisy, as well as changes to the `Drawing` trait to remove the explicit `.into_iter()` call when passing objects to it.